requirements.txt
daphne
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.0
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
billiard==4.1.0
celery==5.3.1
celery-progress==0.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
cron-descriptor==1.4.0
cryptography==41.0.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.10.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-nested-inline==0.4.6
django-phonenumber-field==7.1.0
django-timezone-field==5.1
djangorestframework==3.14.0
drf-yasg==1.21.6
et-xmlfile==1.1.0
executing==1.2.0
frozenlist==1.3.3
gigachat==0.1.8
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
inflection==0.5.1
ipython==8.14.0
jedi==0.18.2
kombu==5.3.1
locust==2.16.1
MarkupPy==1.14
matplotlib-inline==0.1.6
multidict==6.0.4
odfpy==1.4.1
openai==0.27.8
openpyxl==3.1.2
packaging==23.1
pandas
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.16
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.0.1
pydantic_core==2.0.2
Pygments==2.15.1
PyJWT==2.8.0
pypugjs==5.9.12
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.3.1
redis==4.6.0
requests==2.31.0
sentry-sdk==1.27.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.5.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
yandex-s3==0.1.1
yarl==1.9.2
whitenoise==6.5.0
openpyxl
django-filter
channels
djangochannelsrestframework
channels-redis

#oauth
facebook-sdk
google-api-core
google-auth-oauthlib
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos